Removed manual processing of env to rely on pipenv
